--- a/src/utils/astar.rs
+++ b/src/utils/astar.rs
@@ -637,6 +637,7 @@ mod test {
         core::{algebra::Vector3, rand},
         utils::astar::{PathFinder, PathVertex},
     };
+    use std::time::Instant;
 
     #[test]
     fn astar_random_points() {
@@ -759,5 +760,91 @@ mod test {
         assert_eq!(pathfinder.vertex(1).unwrap().neighbours, vec![2, 3]);
         assert_eq!(pathfinder.vertex(2).unwrap().neighbours, vec![1, 3]);
         assert_eq!(pathfinder.vertex(3).unwrap().neighbours, vec![2, 1]);
+    }
+
+    fn generate_grid(size: usize) -> PathFinder {
+        let mut pathfinder = PathFinder::new();
+
+        // Create vertices.
+        let mut vertices = Vec::new();
+        for y in 0..size {
+            for x in 0..size {
+                vertices.push(PathVertex::new(Vector3::new(x as f32, y as f32, 0.0)));
+            }
+        }
+        pathfinder.set_vertices(vertices);
+
+        // Link vertices as grid.
+        for y in 0..(size - 1) {
+            for x in 0..(size - 1) {
+                pathfinder.link_bidirect(y * size + x, y * size + x + 1);
+                pathfinder.link_bidirect(y * size + x, (y + 1) * size + x);
+            }
+        }
+
+        return pathfinder;
+    }
+
+    #[test]
+    fn astar_large_grid_benchmark() {
+        let start_time = Instant::now();
+        let mut path = Vec::new();
+
+        for size in [10, 40, 100, 500] {
+            println!("benchmarking grid size of: {}^2", size);
+            let setup_start_time = Instant::now();
+
+            let mut pathfinder = generate_grid(size);
+
+            let setup_complete_time = Instant::now();
+            println!(
+                "setup in: {:?}",
+                setup_complete_time.duration_since(setup_start_time)
+            );
+
+            for _ in 0..1000 {
+                let sx = rand::thread_rng().gen_range(0..(size - 1));
+                let sy = rand::thread_rng().gen_range(0..(size - 1));
+
+                let tx = rand::thread_rng().gen_range(0..(size - 1));
+                let ty = rand::thread_rng().gen_range(0..(size - 1));
+
+                let from = sy * size + sx;
+                let to = ty * size + tx;
+
+                assert!(pathfinder.build(from, to, &mut path).is_ok());
+                assert!(!path.is_empty());
+
+                if path.len() > 1 {
+                    assert_eq!(
+                        *path.first().unwrap(),
+                        pathfinder.vertex(to).unwrap().position
+                    );
+                    assert_eq!(
+                        *path.last().unwrap(),
+                        pathfinder.vertex(from).unwrap().position
+                    );
+                } else {
+                    let point = *path.first().unwrap();
+                    assert_eq!(point, pathfinder.vertex(to).unwrap().position);
+                    assert_eq!(point, pathfinder.vertex(from).unwrap().position);
+                }
+
+                for pair in path.chunks(2) {
+                    if pair.len() == 2 {
+                        let a = pair[0];
+                        let b = pair[1];
+
+                        assert!(a.metric_distance(&b) <= 2.0f32.sqrt());
+                    }
+                }
+            }
+            println!("paths found in: {:?}", setup_complete_time.elapsed());
+            println!(
+                "Current size complete in: {:?}\n",
+                setup_start_time.elapsed()
+            );
+        }
+        println!("Total time: {:?}\n", start_time.elapsed());
     }
 }