I've noticed significant improvements with this new implementation of A* compared to the current one. I will have to make and run some tests to compare the two versions better. I believe this implementation could see further improvements with reworks to the entire A* (*I left it as is for backward compatibility*). This implementation will also make multithreading easier for both individual searches and parallel searches.

Implements fixes for Issue #442