@TiltedTeapot 2023-10-19
This is my first time writing in Rust so any advice would be helpful!