@TiltedTeapot 2023-10-19
Current results on (windows10 tr2920X):
| Test | Current A* | New A* |
| --- | --- | --- |
| Tests on master | ~0.97s | ~0.33s |