@TiltedTeapot 2023-10-26
When creating a partial path, is the objective to find the most efficient path to the nearest accessible vertex to the objective vertex or the most efficient path to get as close as possible to the objective vertex @mrDIMAS?

In the first case we expect the path to end at the closest available vertex.
In the second case we expect the path to get as close as possible to the objective without wasting energy.

for example the first case would go all the way around the other side of the objective's island if it was closer, whereas the second case would not.