@TiltedTeapot 2023-10-27
Current results on (windows10 tr2920X) single thread:
| Test | Current A* | New A* |
| --- | --- | --- |
| Complete grid (size 10^2) | ~0.049s | ~0.051s |
| Complete grid (size 40^2) | ~0.968s | ~0.277s |
| Complete grid (size 100^2) | ~13.269s | ~1.087s |
| Complete grid (size 500^2) | **>300s** | ~19.037s |
| Island (size 100^2) | **>300s**| ~15.519s |
| backwards travel (size 100^2) | ~19.199s | ~1.899s |

**All benchmark times are total from 1000 iterations*