@TiltedTeapot 2023-11-12
# Final Results (windows10 tr2920X) on single thread:
| Test | Current A* | New A* |
| --- | --- | --- |
| Complete grid (size 10^2) | ~0.049s | ~0.045s |
| Complete grid (size 40^2) | ~1.023s | ~0.194s |
| Complete grid (size 100^2) | ~14.569s | ~0.509s |
| Complete grid (size 500^2) | **>300s** | ~2.822s |
| Island (size 100^2) | **>300s**| ~4.626s |
| backwards travel (size 100^2) | ~21.135s | ~0.781s |

**All benchmark times are total from 1000 iterations*