@TiltedTeapot 2023-11-12
I tried for about a week to get favorable results on multithreading a single search but honestly, unless you are trying to find the best path it is only slower.

The implementation I settled on is significantly faster but does use more memory. I am happy with the result as it is