@TiltedTeapot 2023-11-12
I renamed the path generation methods and separated returning vertex indices and returning positions into different methods instead of using a generic func. These changes make the names clearer and how to use the methods less confusing. I left the `build()` method as depreciated.