@mrDIMAS 2023-11-13
Very nice results! Sorry that I didn't respond, I was busy with other stuff (navmesh path straightening - a.k.a string pulling). About the algorithm itself - I'm I correct that it is still classic A*, but with binary heap that gives major optimization (that basically replaces linear search with O(n) complexity to binary search with O(log(n)) complexity)? I'm a bit confused about the iteration counter (`for i in 0..1000`), why it is needed in the first place? I'm not an expert in A* by any means, so excuse my potentially dumb questions.